+++ provision/scripts/python.sh
@@ -155,29 +155,3 @@ fi
 
 # Add a shortcut symlink to the virtualenv directory in the "ln" directory
 ln -sf "$VENV_DIR" "$APP_DIR/ln/virtualenv"
-
-
-#
-# Install Python dependencies from requirements.txt. If DEBUG is true, also
-# install extra dev dependencies from dev_requirements.txt.
-#
-
-echo " "
-echo " --- Install Python dependencies ---"
-if [[ -f "$SRC_DIR/requirements.txt" ]]; then
-    su - webmaster -c "$VENV_ACTIVATE_CMD && pip install -q -r $SRC_DIR/requirements.txt"
-    echo "Done"
-else
-    echo "None found"
-fi
-
-if [[ "$DEBUG" -eq 1 ]]; then
-    echo " "
-    echo " --- Install Python additional dev dependencies ---"
-    if [[ -f "$SRC_DIR/dev_requirements.txt" ]]; then
-        su - webmaster -c "$VENV_ACTIVATE_CMD && pip install -q -r $SRC_DIR/dev_requirements.txt"
-        echo "Done"
-    else
-        echo "None found"
-    fi
-fi